tick

+++ plugins/dummy/src/main/java/kg/apc/jmeter/dummy/DummyElement.java
@@ -94,7 +94,7 @@ public class DummyElement implements Serializable {
             }
             res.sampleEnd();
         } else {
-            res.setStampAndTime(System.currentTimeMillis(), (long) getResponseTime());
+            res.setStampAndTime(System.currentTimeMillis(), getResponseTime());
         }
 
         return res;
